EliteMobs 9.0.5 - [New] Now using MagmaCore, replacing, improving and synchronizing many core systems. MagmaCore is shaded in, and thus requires no setup from admins. - [New] New system for commands, making EliteMobs safe to reload and able to autosuggest content added during reloads. Almost all commands have also been renamed to follow a planned and predictable structure, and some legacy commands have been dropped - [New] All message logging is now handled by MagmaCore - [New] All configurations are now handled by MagmaCore - [New] Many utility methods are now handled by MagmaCore - [Fix] Fixed issue where wormholes would sometimes be placed in incorrect locations - [Fix] Fixed error where scripts would complain about outdated particles, added automatic particle enum conversion - [New] Unlisted dungeons inside of adventures from the teleports page, those should be accessed through the adventure instead. - [New] Added scale script action, which allows scripters to resize entities. Uses duration to reset the scale to 1, and "scale" for the argument - [Fix] Fixed luck potion effects for default items being incorrectly set to be luck of the sea instead - [Removal] Removed trace system from custom bosses

+++ src/main/java/com/magmaguy/elitemobs/commands/CurrencyCommandsHandler.java
@@ -1,9 +1,9 @@
 package com.magmaguy.elitemobs.commands;
 
-import com.magmaguy.elitemobs.ChatColorConverter;
 import com.magmaguy.elitemobs.config.EconomySettingsConfig;
 import com.magmaguy.elitemobs.economy.EconomyHandler;
 import com.magmaguy.elitemobs.utils.Round;
+import com.magmaguy.magmacore.util.ChatColorConverter;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -81,7 +81,7 @@ public class CurrencyCommandsHandler {
         EconomyHandler.addCurrency(player.getUniqueId(), amount);
     }
 
-    public static void addCommand(CommandSender commandSender, String onlinePlayer, int amount) {
+    public static void addCommand(CommandSender commandSender, String onlinePlayer, double amount) {
         Player player = Bukkit.getPlayer(onlinePlayer);
         if (player == null) {
             commandSender.sendMessage(ChatColorConverter.convert("&8[EliteMobs] &4Player " + onlinePlayer + " &4is not valid!"));
@@ -92,7 +92,7 @@ public class CurrencyCommandsHandler {
         commandSender.sendMessage(ChatColorConverter.convert("&8[EliteMobs] &2They now have " + EconomyHandler.checkCurrency(player.getUniqueId())));
     }
 
-    public static void addAllCommand(CommandSender commandSender, int amount) {
+    public static void addAllCommand(CommandSender commandSender, double amount) {
         for (Player player : Bukkit.getOnlinePlayers())
             addCommand(player, amount);
         commandSender.sendMessage(ChatColorConverter.convert("&8[EliteMobs] &2You have added " + amount + " to all online players."));
